@@ -19,14 +19,14 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "symfony/validator": "^4.0|^5.0",
-        "doctrine/annotations": "^1.6"
+        "php": ">=7.4",
+        "symfony/validator": "^5.0"
     },
     "require-dev": {
+        "doctrine/annotations": "^1.6",
         "bamarni/composer-bin-plugin": "^1.3",
         "captainhook/plugin-composer": "^5.0",
-        "phpunit/phpunit": "^8.0|^9.0",
+        "phpunit/phpunit": "^9.0",
         "mockery/mockery": "^1.0"
     },
     "config": {
@@ -52,9 +52,9 @@
         "psalm_base": "vendor/bin/psalm --set-baseline=psalm-baseline.xml",
         "phpunit": "vendor/bin/phpunit --colors=always",
         "phpunit_clover": "vendor/bin/phpunit --coverage-text --coverage-clover build/logs/clover.xml",
-        "coverage": "vendor/bin/phpunit --coverage-html tests/_reports",
-        "phpcs": "vendor/bin/phpcs --standard=ruleset.xml --extensions=php --cache=.phpcs-cache --colors src tests",
-        "phpcsfix": "vendor/bin/phpcbf --standard=ruleset.xml --extensions=php --cache=.phpcs-cache src tests",
+        "coverage": "XDEBUG_MODE=coverage vendor/bin/phpunit --coverage-html tests/_reports",
+        "phpcs": "vendor/bin/phpcs --standard=ruleset.xml --extensions=php --cache=.phpcs-cache --ignore=src/Cascade.php --colors src tests",
+        "phpcsfix": "vendor/bin/phpcbf --standard=ruleset.xml --extensions=php --cache=.phpcs-cache --ignore=src/Cascade.php src tests",
         "binupdate": "@composer bin all update --ansi",
         "bininstall": "@composer bin all install --ansi"
     }
