@@ -32,11 +32,20 @@ class OrderTest extends \PHPUnit\Framework\TestCase
         $order->invoiceAddress = new Address();
         $order->phoneNumberOnlyAddress = new Address();
 
-        // Leads to six violations in $shippingAddress, none in $invoiceAddress, two in $phoneNumberOnlyAddress
-        $this->assertEquals(8, \count($this->validator->validate($order)));
+        $additionalPHP8Violations = 0;
 
-        // Leads to six violations in $shippingAddress, four in $invoiceAddress, two in $phoneNumberOnlyAddress
-        $this->assertEquals(12, \count($this->validator->validate($order, null, ['Default', 'alternateInvoiceAddress'])));
+        // Additional violations because of the PHP8 attributes - only affects PHP8
+        if (PHP_VERSION_ID >= 80000) {
+            $additionalPHP8Violations = 4;
+        }
+
+        // PHP7: Leads to six violations in $shippingAddress, none in $invoiceAddress, two in $phoneNumberOnlyAddress
+        // PHP8: Leads to eight violations in $shippingAddress, none in $invoiceAddress, four in $phoneNumberOnlyAddress
+        $this->assertEquals(8 + $additionalPHP8Violations, \count($this->validator->validate($order)));
+
+        // PHP7: Leads to six violations in $shippingAddress, four in $invoiceAddress, two in $phoneNumberOnlyAddress
+        // PHP8: Leads to eight violations in $shippingAddress, four in $invoiceAddress, four in $phoneNumberOnlyAddress
+        $this->assertEquals(12 + $additionalPHP8Violations, \count($this->validator->validate($order, null, ['Default', 'alternateInvoiceAddress'])));
 
         // Leads to no violations in $shippingAddress, four in $invoiceAddress, none in $phoneNumberOnlyAddress
         $this->assertEquals(4, \count($this->validator->validate($order, null, ['alternateInvoiceAddress'])));
@@ -47,8 +56,16 @@ class OrderTest extends \PHPUnit\Framework\TestCase
         $order = new Order();
         $order->shippingAddress = null;
         $order->invoiceAddress = null;
+        $order->phoneNumberOnlyAddress = null;
 
-        // Leads to three NotNull violations
-        $this->assertEquals(3, \count($this->validator->validate($order)));
+        $expectedViolations = 3;
+
+        // Additional violations because of the PHP8 attributes - only affects PHP8
+        if (PHP_VERSION_ID >= 80000) {
+            $expectedViolations = 4;
+        }
+
+        // Leads to NotNull violations
+        $this->assertEquals($expectedViolations, \count($this->validator->validate($order)));
     }
 }
